Refactored tests so that all controller test run against both in memory and couchdb stores.
