+++ Fabric.Identity.IntegrationTests/CouchDbApiResourceControllerTests.cs
@@ -1,0 +1,10 @@
+﻿using Fabric.Identity.API;
+
+namespace Fabric.Identity.IntegrationTests
+{
+    public class CouchDbApiResourceControllerTests : ApiResourceControllerTests
+    {
+        public CouchDbApiResourceControllerTests() : base(FabricIdentityConstants.StorageProviders.CouchDb)
+        { }
+    }
+}
