+++ Fabric.Identity.IntegrationTests/CouchDbClientRegistrationTests.cs
@@ -1,0 +1,10 @@
+﻿using Fabric.Identity.API;
+
+namespace Fabric.Identity.IntegrationTests
+{
+    public class CouchDbClientRegistrationTests : ClientRegistrationTests
+    {
+        public CouchDbClientRegistrationTests() : base(FabricIdentityConstants.StorageProviders.CouchDb)
+        { }
+    }
+}
