+++ Fabric.Identity.IntegrationTests/CouchDbIdentityProvidersControllerTests.cs
@@ -1,0 +1,10 @@
+﻿using Fabric.Identity.API;
+
+namespace Fabric.Identity.IntegrationTests
+{
+    public class CouchDbIdentityProvidersControllerTests : IdentityProvidersControllerTests
+    {
+        public CouchDbIdentityProvidersControllerTests() : base(FabricIdentityConstants.StorageProviders.CouchDb)
+        { }
+    }
+}
