+++ Fabric.Identity.IntegrationTests/CouchDbIdentityResourceControllerTests.cs
@@ -1,0 +1,10 @@
+﻿using Fabric.Identity.API;
+
+namespace Fabric.Identity.IntegrationTests
+{
+    public class CouchDbIdentityResourceControllerTests : IdentityResourceControllerTests
+    {
+        public CouchDbIdentityResourceControllerTests() : base(FabricIdentityConstants.StorageProviders.CouchDb)
+        { }
+    }
+}
