+++ Fabric.Identity.IntegrationTests/IdentityProvidersControllerTests.cs
@@ -2,6 +2,7 @@
 using System.Net;
 using System.Net.Http;
 using System.Threading.Tasks;
+using Fabric.Identity.API;
 using Fabric.Identity.API.Models;
 using Newtonsoft.Json;
 using Xunit;
@@ -11,6 +12,10 @@ namespace Fabric.Identity.IntegrationTests
     public class IdentityProvidersControllerTests : IntegrationTestsFixture
     {
         private readonly string _identityProvidersBaseUrl = "/api/identityProviders";
+
+        public IdentityProvidersControllerTests(string provider = FabricIdentityConstants.StorageProviders.InMemory): base(provider)
+        { }
+
         [Fact]
         public async Task IdentityProvidersController_Get_ReturnsProviders()
         {
